requirements.txt
asgiref==3.3.1
Django==3.1.5
django-extensions==3.1.3
django-filter==2.4.0
psycopg2==2.8.6
pydotplus==2.0.2
pyparsing==2.4.7
pytz==2020.5
sqlparse==0.4.1
